requirements.txt
requests~=2.31.0
flask~=2.2.5
WTForms~=3.0.1
python-dotenv~=0.21.1
flask_bootstrap
flask_sqlalchemy
flask_wtf
bs4~=0.0.1
IMDbPY
beautifulsoup4~=4.12.2